MXS-3742 Conditional rendering table rows based on visibility of headers
There are several issues when headers and rows are filtered out instead of
being hidden. e.g. When columns visibility are filtered out by toggling the
columns, deleting the row from the collection becomes impossible as the row's
value at that column is also filtered out. Since 2d array is being use to render
table rows, there is no unique key to detect which row is being deleted, so
the entire row array needs to be compare which is not possible when row's value
is filtered out.

This adds `hidden` property to the header so when toggling the visibility of
a header, in each row, at that column, the cell won't be rendered.
This alsos do minor refactor on grouping feature to adapt this new `hidden`
property.