Fix longblob and mxs1110_16mb
The payload calculation was wrong and ended up with chunks eight times
larger than was intended. This in turn caused the execution of those
statements to fail as the value configured for max_allowed_packet was
about ten times smaller than was actually needed.

Configuring max_allowed_packet to 160MB would've fixed the test but it
caused the longblob test to take about a minute. With the correct chunk
size calculations, it takes less than ten seconds.