MXS-3583 Ensure all $all tests pass
With these changes all tests for $all pass. The challenge
is that if $elemMatch is used in the array of $all, then
it will be encountered at a stage when it no longer is
possible to just generate the relevat code. To avoid a
two-pass approach the $elemMatch items are stored and
processed later.

Further, the way a bson element is converted to a string
depends to some extent on the context where it will be used;
as an argument to be gived to a JSON function or as a value
to be compared with a result returned by a JSON function.
Thus, when converting a value you must specifiy the purpose.

This boils down to the difference between
JSON_CONTAINS(... X ...) and JSON_VALUE(...) = X.