Remove unstable tests
They've been unstable for years and are either not good tests or not much
use in their current form.

Removed the unstable label from longblob now that it no longer tries to
insert gigabytes of data. This leaves only the xpandmon test and the
gssapi tests with the UNSTABLE label, the latter of which is being
rewritten to be much smaller.