MXS-3759: Fix unexpected error handling
If a complete resultset is received along with an error that signals
server shutdown, the trailing error packet must not be removed if the
client has seen any part of the result. Doing so would cause the client to
hang as it expected an EOF packet that would never arrive as the server
substituted it with an ERR packet which readwritesplit discarded.

In addition to this, if the client hasn't seen the result, the whole
result must be discarded. This will prevent partial results from being
sent to the client when we know we won't be receiving the rest.

It also turns out that the server never sends any errors when a connection
is killed via the KILL command. This means all unexpected errors that
readwritesplit intercepts are caused by a server shutdown.