MXS-3745 Handle clearing persistent storage
When storing history/favorite queries to localStorage, the maximum storage size
can not be determined easily as different browsers have different size and user
can even increase the size.

So, the most appropriate way is to show a snackbar indicating error
and inform user to delete some to create more space.