Fix pinloki_galera
With MariaDB 10.5, the gtid_domain_id must not be set for Galera nodes if
we want the binlog contents to be consistent on all nodes. Since no other
test relies on this and versions older than 10.5 don't really work with
GTIDs, it should be safe to change it to the same value on all nodes.

Changed the hard-coded sleeps to a 30 second loop. This should help both
speed up the test and help with cases where the test fails due to
replication lag. This seemed to happen a few times even when everything
was running locally.