Fix auto truncating cell text in `data-table`
When a cell shows toggle button, the function handles showing tooltip
for truncated text should subtracting toggle button's width from
wrapper's width in order to detect whether the cell value is truncated or
not.

This also replaces `width` attribute of `th` with inline style as
`width` attribute is not supported by HTML5.