Fix charsets in mysqlmon_fail_switch_events
The ALTER EVENT changed the character set back to the system default which
isn't always latin1. The test should also not expect the database charset
to be any pre-defined one as nothing in the test defines it. It isn't
really relevant to test as the monitor is not expected to or even capable
of changing it.