Add `expandAll` props to `data-table`
Adding this props to make statistics table in server details page
to be rendered as tree data table as it contains `response_time_distribution`
and has a lot of nested childrens. Hence, expanding all nodes by default for this
table is not appropriate.

This adds `expandAll` props to components using `data-table` and having
`isTree` props.

This also cleans up unnecessary props in components using `data-table`.