Combine rw_select_insert into sanity_check
The use of the COM_ counters for detecting where a query is routed is more
fragile than storing the server ID in the table and then using it to
compare where it ended up.

Although a plain INSERT isn't directly tested, the following SELECT will
confirm where the query ended up going. If the test is run using MariaDB
10.5, we can use INSERT ... RETURNING to verify that the INSERT itself
ended up on the right server.