Fix truncated menu position
If a cell is not supposed to have its value truncated,
it shouldn't be wrapped with a inline element. This causes
offset top issue in v-menu when using custom cell slot as the
height of the slot element is not occupied to its maximum height.

This renders the wrapper inline element only when `header.cellTruncated` is true
and add `offset-y` props to v-menu to auto align the menu instead of using `truncatedMenu.y`