MXS-3659 Detect whether cell value is truncated smarter
This uses `doubleRAF` to check `scrollWidth` and `clientWidth`
of the truncated element. This eliminates the need for using inline
element since it causes slot element to be render as an inline element.

This is a preliminary commit to make custom slot controls its padding
instead of setting padding on `td`. When padding is set on `td`,
overriding cell events such as mouseover, mouseleave or mouseenter
become difficult.