MXS-3766 Exclude INSERTs from function usage prevention
An insert like "INSERT INTO t (f) VALUES ...", where the column ("f")
is mentioned explicitly, will be reported to be using the column "f" in
conjunction with the function "=", which will cause it to be rejected if
"f" is a column that should be masked.

So, INSERTs need to be excluded from the check. That does not effectively
change anything, as an INSERT without named columns will not ever be
rejected, which would be nonsensical anyway.

The check cannot be limited to just SELECTs, because that would
allow you to probe a value e.g. using UPDATEs.