MXS-3711: Make qlafilter modifiable at runtime
All qlafilter parameters can now be updated at runtime. The parameter
updates behave the same way the other filters do: a new configuration is
taken into use only by sessions created after the change took place.

This revealed some problems that already existed in the filter: if a log
file already exists but it was created with different values of log_data,
new entries will use the new log_data format which will cause the data and
the header information to be out of sync. This means the CSV format isn't
reliable if the same filename is used.