Fix COM_STMT_EXECUTE metadata injection
The metadata injection should only be done if the prepared statement has
parameters in it. If it has none, the pointer would point at the
past-the-end pointer in the buffer resulting in undefined behavior. This
also caused warnings about malformed COM_STMT_EXECUTE packets to be
logged.

This only affects code that's in the upcoming 6.2 release and thus does
not need to be backported.