Add `miragejs` to use as an API mocking library
API data was mocked with hard-coded variable stored in `tests/unit/utils.js` which
is inefficient and hard to maintain.

This adds `miragejs` to use as an API mocking library, this removes the
need for stubbing axios requests and improves test code readability.

At the moment, only `Servers.spec.js` is refactored to use this API mocking.