Make mxs559_block_master less intense
As it was, a normal run of the test caused about 80k lines of errors which
was about 85% of the whole test job output.

In addition, the test easily caused problems for other tests by exhausting
the local TCP ports by repeatedly opening new connections in a while loop
with no sleeps. Waiting for even a second should reduce the number of
created TCP connections to an acceptable level.