Fix RegexValue regression
The set_options did nothing as it didn't recompile the pattern which is
why it was removed in 2.5. Presumably due to a merge gone bad, this code
remained in 6.1 but was removed from 6.

Also fixed other uses of set_options to recompile the pattern.

This commit fixes the ccrfilter_test test case.