MXS-3353: Add sync option to tee filter
The sync option enables a synchronous mode which waits for all replies to
complete before new ones are allowed. This gives a guarantee that all but
the latest query will have successfully completed when the client reads
the result of the last query. In practice this means that if a session
closes in a controlled manner, everything will have been written and read
back. This relies on the fact that the last packet the client will send is
the COM_QUIT and before that is routed, all replies must have returned.

The major limitation of this mode is that any errors in the branch target
will cause the whole session to be closed. Given that the goal of this
mode is to synchronize two databases, it should be an acceptable one under
most circumstances.