MXS-3353: Enable sync mode test case
The old test would not pass with the fixed code as the INSERT would block
due to the LOCK TABLE that was done before it.

Renamed objects in the test config to newer names and enabled log_info.
The test isn't very long and having log_info speeds up the process of
analyzing test failures.