MXS-3583 Add support for searching using DBPointer
A DBPointer must be created from an ObjectId, yet when a lookup is made
using it, the id is passed as a string without any information that would
reveal that the source is a DBPointer. Thus, if an _id lookup is made and
the provided id is a string which is 24 characters long and a valid hex
number, we cover the case that it might be an ObjectId in disguise.