MXS-3793: Fix GCUpdater shutdown and startup
The modification of m_running must be done on the thread that starts and
stops the thread. This guarantees that the state is always correct and the
thread in the GCUpdater::run function only reads it.

A std::future was used by both the smartrouter and the qlafilter to manage
the thread that runs the GCUpdater::run function. This is not required and
it can be made into an internal implementation detail of GCUpdated without
changing the external functionality of it.