MXS-3798: Fix user account cache sync
Making the updater thread wait for the RoutingWorkers prevents the Service
from being destroyed while the message sits in the RoutingWorker's message
queue. This works as the MainWorker waits for the updater thread in
Service::~Service. Waiting on the RoutingWorkers will not cause a deadlock
as they never wait on other threads.