MXS-3796: Add test case
Added a test case that triggers a crash in MaxScale. This isn't the exact
same situation that happened in the original bug but it should result in a
hang if it doesn't crash.

The reason why this happens is that the buffer passed to SetParser::check
in MariaDBClientConnection::handle_variables points to a local variable
and the call to gwbuf_make_contiguous inside SetParser::check causes the
old buffer to be first consumed into an empty buffer and then freed. Since
the new buffer is stored at a separate address, the mxs::Buffer that
originally contained it will now point to freed memory.

If accessing the freed memory does not immediately cause a segmentation
fault, the buffer would appear to be empty and would end up causing the
hang that is described in the issue.