MXS-3796: Fix buffer handling in handle_variables
The client protocol code was changed to use a mxs::Buffer for version
6. The variable handling code could end up reallocating the buffer which
wasn't correctly synced with the owning mxs::Buffer object.

Eventually, the variable handling code should be rewritten to use a
mxs::Buffer instead. This would also allow iterators to be used which
would make it possible to parse the buffer without making it contiguous.