MXS-3790: Export liblua.so symbols in luafilter
Using RTLD_GLOBAL causes libraries loaded with dlopen to behave as if they
were linked against the program loading them. That is, the symbols from
the loaded libraries are made globally visible to all subsequent libraries
loaded with dlopen. As lua uses dlopen to load modules, this fixes the
problem of some luarocks libraries not working with MaxScale.