Fix table tooltip and context menu position
Even `id` of each node in schema tree is unique, it can not always
be used as an id attribute for html element as it isn't valid.
For example: if table/column/database name contains space or any
characters against html valid values for `id` attribute, the tooltip
and context menu won't be visibled at expect position.

This adds unique `key` attribute to each node in the schemas tree.