Treat xpand group changes as transient errors
This keeps the backends in use even if they reject the
authentication. Normally when a server rejects the authentication it is no
longer used by the routers. This is done to prevent repeatedly
reconnecting to the same server that doesn't have the required user
account that the other servers have.