MXS-3803: Improve Trx::add_stmt logging
The MXS_INFO call should be unconditionally made as there are cases where
a message must be logged even if the log level for it isn't enabled. This
is automatically handled by the macro and the arguments are only evaluated
whenever logging is required.

Added an extra debug assertion that makes sure the target is non-null and
added a message to the target comparison assertion.