MXS-3803: Ignore hints with transaction_replay
If a routing hint would cause a query to be routed to a server that's
outside of the current transaction, it must be ignored if
transaction_replay is enabled. Not doing so breaks the internal logic and
causes debug assertions.

Technically this would be acceptable but in hindsight the decision to
allow hints to break the transaction routing logic was a mistake: rarely,
if ever, a `SELECT ... FOR UPDATE` that happens to match a
namedserverfilter rule is truly intended to be routed to a server that's
not the target of the transaction.

The correct thing to do would be to ignore all routing hints inside
transactions but this affects all routers and can possibly prevent some
semi-valid use-cases where statements that are known to be safe. Since the
use of the namedserverfilter to act as a user-configurable override of the
routing logic is fairly common, only changing this for transaction_replay
fixes the problem but prevents existing setups from failing.

Added a test case that verifies that transaction_replay causes routing
hints to be ignored. Also documented the new behavior in the
readwritesplit documentation.