MXS-3583 Add more information to Path::Incarnation
Now, a provided MongoDB search path is first parsed and a forest
of possible interpretations is created. Then, from the leaf nodes
of the forest the corresponding JSON paths are created.

The main problem is that given a MongoDB query "a.b", "b" can
refer to the "b" field of the document at "a"
- '{ "a": { "b": 42 } }' - or to the "b" field of a document in the
array at "a" - '{ "a": [ { "b": 42 } ] }' -, which correspond to the
JSON paths "$.a.b" and "$.a[*].b" respectively. Further, when the
latter path is used, in some contexts, in order to obtain the correc
result, it must be ensured that "$.a" is an array.

Another similar problem is that given a MongoDB query "a.1", "1" can
refer to the "1" field of the document at "a" - '{ "a": { "1": 42 } }' -
or to the 1st element of the array at "a" - '{ "a": [ 42 ] }'.

And the query can be more than just two components; e.g. "a.b.c"
where "a", "b" or "a" and "b" can be arrays.