MXS-3802 Don't always show scrollbar in `virtual-scroll-table`
Always displaying scrollbar makes the UI looks clumsy. The scrollbar
should only be added when needed.

Since the scrollbar is not overlay added, the header has already had
`dummy-header` element to always take space.
So this adds `dummy-cell` element so that when table body is not vertically
overflowed, it's rendered to take space.