Return assigned GTID if no events have been replicated
In MariaDB the value of @@gtid_slave_pos is a global variable which means
that once it's set, the same value is returned even if replication has not
started. To retain this same behavior, the binlogrouter should return the
assigned GTID if a real GTID has not been read yet.