Update MaxCtrl dependencies
NPM reported vulnerabilities, fixed them with `npm audit fix`. The globby
package started to cause problems after this update due to it being a
ECMAScript module which pkg did not understand. Removing it seems to fix
it and since it was added to 2.3, it's most likely already fixed in the
package that originally required it.