MXS-3514 Fix an error in the test and a bug in the dectection
The test was only testing the regular expression part. Added
detect_special_query(). There was a query starting with
"--line", which is invalid sql, changed to "-- line".

The detection's prefix check did not take into account a
variable number of spaces. Since the detection need not be
perfect, but preferably very fast, it now only checks for the
3 first letters for "USE", "KIL" and "SET". Only "SET" can lead
to uneccesary regex matching, but likely appears with such
low frequency that it does not matter.