MXS-3802 Add `rowData` as data property of cell slot
`rowData` stores the reference to the entire row values. This data
is useful in case of modifying a cell value requiring to check
dependent cell value. e.g. In `alter-cols-opts` component,
changing the data type of a column, the charset/collation cell input
needs to check that column type in order to handle disable the input.

This also fixes `vertical-row` cell slot.