MXS-3813: Support PS direct execution in hintfilter
The tracking is relatively simple: if a binary protocol PS is received
with the ID -1, it is substituted with the ID of the previous statement
when searching for the hint. This can be done completely asynchronously
with regards to the execution of the PS.

The test case reproduces the problem and verifies that the new code solves
it.