MXS-3583 Add custom converter from JSON to BSON
If the JSON obtained from the database contains data that looks
like a particular BSON type, without containing everything that
is needed or containing everything but of the wrong type, the
default conversion will fail with an exception. If that occurs,
we create the BSON object manually. While doing that, we convert
a document that fulfills the requirements of an ObjectId into one.