MXS-3634 Add `allowEnterToSubmit` props
`base-dialog` is configured to listen on `Enter` key press to submit
form. Though this brings convenience in most cases, but when dialog
contains extra inputs that doesn't need to be validated, this cause
form to be submitted. e.g. `query-editor` renders in dialog.

This adds `allowEnterToSubmit` to handle listen on `Enter` key
press event.