MXS-3802 Make `codeFormatter` as a helper
`query-editor` automatically format sql code when
component is mounted. After that, it won't format
until user press shortcut key or right click to open
context menu and choose `Format Document`. This is
default behavior but when the sql code is auto generated
in `DDL_EDITOR` mode, there is a need to re-format
sql code before showing the editor.

This makes it as a helper function so that it can be reused.