MXS-3782: Fix autocommit tracking
When session_track_trx_state is used to track the transaction state, the
implicit transaction started by SET AUTOCOMMIT=0 was not detected as the
transaction characteristic tracking only tracks explicit transactions.

Using the server status from the OK and EOF packets allows a failsafe for
this which can still be overridden by the more finer grained transaction
tracking.