MXS-3802 Fix incorrect place of comma in `handleBuildColDfnSQL`
Adding comma except first iteration is inaccurate in an edge
case where the first updated column only has changes related to `PK`
or `UQ`. As a result, checking for first index in order to not add comma
is incorrect.

This filters `updatedCols` by keeping only columns having column
definition changed. The downside of this approach is to have to
iterate through `updatedCols` one more time.