MXS-3802 Remove `cell-content--vert-center` wrapper class
`vertical-align: middle` and `display: table-cell` approach
doesn't work properly in most browsers. Apart from that, the table
is intended to render text and text is aligned vertically using line-height
attribute. In order to use custom slot to render input element with
vertical aligned, the input has to be an inline or inline-block element.
Alternatively, vertical align can be done using flex box but it has
to be done in the parent component as `virtual-scroll-table` has
text overlapped issue when filtering table row.

This removes `cell-content--vert-center` and add a flex wrapper
element around `column-input` to align input vertically.