MXS-3802 Refactor `buildColsDfnSQL` to make it reusable
When `ADD COLUMN`, it has similar syntax except `CHANGE` keyword
and old column name. Hence, this function can be resuable
by adding `isChanging` boolean flag.  If `isChanging`
is true, each `col` in `cols` array will be a diff col that is
to have `newObj` and `oriObj` property. Otherwise, `col`
will be an object with all keys similar to `newObj` which
can be then used to `ADD COLUMN`.