MXS-3802 Use `@input` instead of `@change` event
After changing the content of the input, `change` event will be
emitted only when the element loses focus. This leads to the situation
where user finishes changing the value of a cell then quickly moves the
cursor to the `Apply` button, it requires two clicks to click the button.
One click to lose focus from the input, the other click to click he button.

This switches to use `input` event as it emits immediately.
Except `v-checkbox` as it only has `change` event and it has similar
behaviour as other types of inputs.