MXS-3802 Capitalize columns alter table header
The headers return from the query is in lowercase and have
underscore between words. This helps to destructuring object
easier. However, it makes the UI look inconsistent as other tables
are having uppercase headers. But having uppercase header would
increase the width of the table. However, having table header capitalized
is acceptable.

This adds `capitalize` property to header. For headers having
underscore, this uses header slot to customize it.