MXS-3834: Add reconnect endpoint to SQL API
This makes it possible to reconnect an existing connection that failed due
to some network error.

This often happens when a query times out and would previously require the
creation of a new connection. Due to the fact that the GUI doesn't store
the credentials, this would mean that the user had to first disconnect the
current connection and then re-enter all the information for the same
connection.