Fix connection dialog form for not clearing accurately
`timeout` value is cleared when open the dialog then close it.

`base-dialog` is configured to auto clear form inputs when dialog
is closed. So the watcher to check the dialog is closed, in order
to reset to initial state needs to be run in the next tick so
that form value can be reset to initial state accurately.