MXS-3802 Add `Generated` column input
This input allows to alter column to use generated (Virtual and Persistent/Stored)
The expression input is now combined with `default` column to become
`Default/Expression` column. If `Generated` input cell value equals
to `(none)` then the value in the `Default/Expression` cell will be the value
for `DEFAULT`, otherwise it's the expression value.