MXS-3802 Minimize columns in alter columns table
Two many input columns narrow the width of each input in small screen
resolution. To deal with this, charset, collation, and comment inputs are
removed from the table. There should be a dropdown input below the table
to select a column name to reveal additional inputs.

This commit only removes those inputs from the table. The following
commits will handle showing it below the table.