MXS-3802 Recalculate table body's height of `virtual-scroll-table`
`virtual-scroll-table` uses value of `height` props minus header's height
to assign as the height of table body. However, the value of `height` props
and number of rows are dynamic, using `height` as table body's height is not
accurate. This leads to the situation when there is a need to render some elements
below the table but there is no space left as the table has already occupied it.

This renames `height` to `maxHeight` and recalculate table body's height
smarter. It compares the actual height needed to maximum table body's height
to get accurate height.