MXS-3583 Return write errors individually in writeErrors
Write errors, both (1) such related to the command being invalid
or untranslatable (into SQL) and (2) such related to the failed
execution, should be returned as separate errors in a writeErrors
array.

Now that is done for (1) as well. However, it should still be
modified so that the whole execution is not aborted, but that
successfully translated commands should be executed.