MXS-3583 Refactor handling of update operations
Clearer if every operation is handled by a separate function.
However, this is broken since several update operators should
be handled in a sequence so that the result of one is used as
the input for the next. Currently, using multiple operators
causes incorrect SQL to be generated. To be fixed.