MXS-3802 Undo the removal of charset, collation, comment inputs from table
Undo it as receiving feedback from Irina. Having those inputs below
the table can cause confusion to users. A better approach would be
to have a button/icon to show extra inputs. Small screen resolution should
be detected to hide those inputs by default. The implementation of this
will be done in the next commits.

This also moves those mentioned inputs to the end of the table.