MXS-1350: Add XA transaction parsing
The qc_sqlite parser fully parses the two relevant XA START and XA END
statements and treats them as if they were a normal transaction.

The TrxBoundaryParser class that is used when no query classification is
done only partially parses the needed XA statements. This can cause false
positives when SQL with syntax errors is used. Adding full support for it
requires more changes than the basic functionality does which is why it's
better done in a separate commit (or possibly not at all if we deem the
risk to be acceptable).

The XA transactions were already tested by the query classifier unit tests
and also pass with these additions now that both the query classifier and
the TrxBoundaryParser support it.