Use std::vector for the transaction buffer
This drastically improves the performance with large transactions. The act
of calculating the buffer size ended up creating and destroying a string
which caused the current contents of the std::ostringstream buffer to be
copied.

With this commit, a 400MB transaction takes around 20 seconds whereas
before it didn't finish in five minutes.