MXS-3769: Track SET TRANSACTION state
The relevant forms of SET TRANSACTION are now tracked. When a client
changes the access mode with SET TRANSACTION, only the next transaction
will use the specified access mode. If the SESSION keyword is used, it is
set as the default value for all future transactions. This is

The new QUERY_OP_SET_TRANSACTION value isn't actually used to detect the
SET TRANSACTION statement as it is only available via the query classifier
and the query classification is not always done. However, having it helps
make sure that it's not confused with a normal SET statement.