Fix duplicated completion items
When having two instances of monaco editors on the same page, completion
items are duplicated. The root cause of this behavior comes from the fact
that monaco editor uses a service web worker for a language to save memory
and `registerCompletionItemProvider` is additive not replaceable.

This adds `skipRegCompleters` to conditionally skip `regCompleters`.
In the `execute-sql-dialog`, `skipRegCompleters` value is determined by
using the current editor mode .i.e. when there are more than 2 monaco editor
instances are being visible on the same page, it returns true.