MXS-3850: Ignore SET GLOBAL statements
They do not change the state of the current session, only new sessions
created after it.

The syntax for the SET command remains the same but now the actual
variable processing is done by maxscaleSetVariable instead of the
maxscaleSet function. This commit does not yet remove the ExprList that is
constructed and passed to maxscaleSet as removing it would be somewhat
complex and is better done in a separate commit.

Modified the unit tests to correctly expect the lack of transaction bits
in the cases where the GLOBAL keyword is used.