Standardize input height
Since the change of `Fix input padding style issue`, all inputs need to
specify the height value. This is because vuetify's input component accept
height value as props but the default min-height is hard-coded with sass variable.
So the height value can not be smaller than that hard-coded value. To workaround this,
`v-input__slot` min-height was set to 0 in order to make the `height` props of all
inputs to accept value smaller than the hard-coded sass
variable.

This uses `36` as a standard deviation value for all inputs according to
Mariadb Style Guide except some inputs in the query editor.