MXS-3815 Stop file scan on first GTID_LIST_EVENT
While searching for the initial GTID requested by a client
files are scanned in reverse order as it is most likely that
the GTID is in one of the latest files. There is no reason
to continue to scan a file after the GTID_LIST_EVENT
indicates the GTID is in a prior file, instead immediately
move to the next file to scan.