MXS-3856: Return an error for impossible causal reads
When a causal read cannot be performed due to the session being inside of
a read-only transaction, an error is returned. This lets the client know
that something went wrong but it could be tried again.

The exact error that is returned when a causal read fails is also included
in the documentation for `causal_reads`.