MXS-3583 Use _id value in a query only if it is valid
When upserting, the document is build from the query object
and or the update object. However, a possible _id in the query
object must only be used if it is a valid value (and not a
query itself). If it is not valid to be used as an _id, then
a proper _id should silently be generated.