MXS-3583 Add initial support for command Explain
Explain is conceptually similar to DESCRIBE. However, it is
used in many tests, not to find out how things would go, but
to verify that the server behaves exactly as expected. That is,
it's the server and not the protocol behaviour that is tested.

As these tests are in between other tests, the aim is return
data that mimics the MongoDB behaviour.