MXS-3836: Fix refresh_databases
The parameter once again causes the databases to be refreshed, if
possible, whenever a USE statement fails due to an unknown database.

As the parameter has not worked for a long time, it is now disabled by
default. This will prevent unwanted changes in both performance and
behavior that can now happen.