Make table mapping more efficient in certain cases
If duplicates are being ignored, there's no need to perform the shard
location lookup. Due to the inefficient way the location lookups are
implemented (lots of temporaries being constructed), this is a measurable
improvement in the database mapping performance.

The check for the dot in the data can also be removed as the SQL always
includes it.