MXS-3880: Remove bias value for unconnected servers
Removing the bias provides better distribution of the open connections
when a server restart forces all clients to perform a reconnection. With
the old biased score the steady-state distribution of requests between two
candidates ended up somewhere around a 40/60 ratio favoring the open
connection. This caused the open connection to eventually receive more
traffic than the other one even if it could process more.

Since the workloads that suffer from this are the ones that would benefit
the most from no bias value at all, removing it is most likely the best
option.