MXS-3883: Fix listener nested parameter serialization
As nested parameters are only used by the nosqlprotocol, this problem can
be solved for 6.1 by storing the nested protocol parameters
separately. This solves the immediate problem but it isn't a very good
solution.

The proper way to sovle this would be to extend the protocol module API to
behave like all other modules do: declare a getConfiguration endpoint and
use that to configure and persist the modules. The problem is that this
requires a lot of changes to the APIs and some extra code in all protocol
modules.