MXS-3633 Simplify text options in the context menu
Text options such as `Place Qualified Name In Editor (Quoted)`,
`Place Qualified Name in Editor`, `Place Name in Editor (Quoted)`
can be shortened as those options are already sub-options of
`Place to Editor` option.

By shortening the text, the handler functions for inserting text to
the editor and copying text to the clipboard can be combined. This
also reduces repetitive code.