MXS-3881: Fix LDLI-in-trx handling in readwritesplit
If transaction_replay was enabled and a LOAD DATA LOCAL INFILE was done
inside of a transaction, the state tracking would not pick the branch that
allowed streaming results to the master. Normally the whole LOAD DATA
LOCAL INFILE would be streamed to the master.

This was caused by a wrong assumption that the LOAD_DATA state would only
be in the ACTIVE state when it can be both ACTIVE and END during the
routing of a LOAD DATA LOCAL INFILE.