MXS-3583 Change how $inc and $mul are generated
The prototype of JSON_SET is

    JSON_SET(doc, path, value [, path, value]*)

which works fine if the path is non-scoped. However, if the path
is scoped, e.g. "a.b", then you must do things differently depending
on whether "a" exists or not. If it does not exist, you need to build
a document hierarchy. However, 'path' above cannot be a MariaDB
IF-function, but the IF must be the outer function and the
JSON_SET the inner. And since there can be multiple paths, in
order to enable IFs (next commit), the setting must be done
recursively like

    JSON_SET(JSON_SET(doc, path2, value2), path1, value1);

When done like this, every single JSON_SET can be

    IF(JSON_QUERY(...), JSON_SET(arguments1), JSON_SET(arguments2))