Log test output into MaxScale log
This should help detect whether the test failures are related to the
monitor not picking up the unblocked node or whether it is something
else. Regardless of the test failure, having the corresponding messages in
the MaxScale log helps keep track at which point of the test each message
was logged.