MXS-2353: Add service level log configuration
Due to the way the logging system works, the service level configuration
cannot truly override the global configuration. However, this allows
simple booleans to be used which would not be possible with the
three-valued logic (on, off, default) required by the overriding version.

As the service level logs are inspected on a rather frequent basis, they
should be stored in a way that is fast to access. The fastest way to do
this would be to store it in the Session as the data is already stored
there. The problem with this is that freezing the logging configuration at
the session creation time by copying it from the service can cause
problems if a long-running session ends up filling the logs with
messages. For this reason the parameter is kept "global".