Wait longer in mxs559_block_master
The monitor seems to sometimes take two ticks to notice that the master
server is available.

This could be explained by the monitor already being between ticks when
the thread that waits for the tick to increment fetches the old value of
ticks. Then when the monitor updates the ticks to the new value, the state
of the server is not what is expected as it was a "partial" monitoring
loop. This means that waiting for two ticks should always guarantee that
at least one full monitoring loop takes place.