MXS-3544 Install `vue-virtual-scroll-list` library
The library is needed in order to have virtual log lines because
`v-virtual-scroll` from vuetify hasn't supported dynamic height
yet. The library is chosen over other libraries such as `vue-virtual-scroller`
because it uses unique id as a key to rerender items instead of index,
so that new items can be prepended without rerendering the whole list.

https://github.com/vuetifyjs/vuetify/issues/11755