MXS-3893: Fix connection_timeout idle calculation
The idle time calculation used only the last network read when both
networks reads and writes should be counted. If we write data to the
client it means it hasn't been idle as it was doing something and we wrote
the reply to it to the client.

The case where the connection was truly idle for a time longer than the
configured connection_timeout did not check if the connection is logically
idle. This caused long-running statements to fail if they took longer than
the configured upper limit. By checking that the session is idle, we
prevent this and as the network writes are now taken into account in the
idle time calculation, we also prevent the case where a client would get
immediately disconnected after it received the result of a long-running
query.

Removed the misleading comment from the is_idle() method as it referred to
the old use of the is_movable() method.