MXS-3544 Move `chosenLogLevels` component's state to vuex module state
This creates `chosen_log_levels` and its mutation in `maxscale`
vuex module. This also uses `&priority` params to filter logs.
Since the fix of the memory leak issue is done by using `virtual-list`,
filtering logs need to be refactored as well. So this commit
is a part of https://jira.mariadb.org/browse/MXS-3620.